combine evidence methods into single file set
